updated distinction between score/value
The new distinction is much more hardline: value is just for the actual 
value derived by the function. Score is now the only way to get the loss 
rescaled between 0-1 (unless the value happens to be this).

 Combination losses are updated to reflect this, and use loss.score() 
instead of loss.value, hopefully caught all of the ones where I needed 
to invert min/max or change default invert value.

TODO, many of the losses would be better as configurable logistic 
functions about some mean with a configurable linear range